+++ integration-tests/code-generator-tests/src/main/java/com/squareup/anvil/test/Trigger.kt
@@ -7,3 +7,6 @@ class AnyClass
 
 @Trigger
 fun abc() = Unit
+
+@Trigger
+val def = Unit