+++ integration-tests/code-generator-tests/src/test/java/com/squareup/anvil/test/GeneratedCodeTest.kt
@@ -84,6 +84,12 @@ class GeneratedCodeTest {
     assertThat(generatedClass).isNotNull()
   }
 
+  @Test fun `the property class is generated`() {
+    val generatedClass =
+      Class.forName("generated.test.com.squareup.anvil.test.GeneratedPropertyClass")
+    assertThat(generatedClass).isNotNull()
+  }
+
   @MergeComponent(Unit::class)
   interface AppComponent {
     fun otherClass(): OtherClass